+++ types/crypto.go
@@ -1,0 +1,11 @@
+package types
+
+import "github.com/ginchuco/ginchu/crypto"
+
+func MerkleTree(items [][]byte) (root []byte, tree [][]byte, err ErrorI) {
+	root, tree, er := crypto.MerkleTree(items)
+	if er != nil {
+		return nil, nil, ErrMerkleTree(er)
+	}
+	return
+}